Strip out old HF code
