+++ pixeltable/func/__init__.py
@@ -2,6 +2,5 @@ from .signature import Signature, Parameter
 from .function import Function
 from .external_function import ExternalFunction, ExplicitExternalFunction
 from .nos_function import NOSFunction
-from .huggingface_function import SentenceTransformerFunction, CrossEncoderFunction, ClipFunction
 from .function_registry import FunctionRegistry
 from .util import udf, make_function, make_aggregate_function, make_library_function, make_library_aggregate_function, Batch
