+++ pixeltable/tests/test_functions.py
@@ -161,15 +161,15 @@ class TestFunctions:
         assert status.num_excs == 0
 
         # run multiple models one at a time in order to exercise batching
-        from pixeltable.functions.huggingface import sentence_transformer_2, sentence_transformer_list_2
+        from pixeltable.functions.huggingface import sentence_transformer, sentence_transformer_list
         model_ids = ['sentence-transformers/all-mpnet-base-v2', 'BAAI/bge-reranker-base']
         num_dims = [768, 768]
         for idx, model_id in enumerate(model_ids):
             col_name = f'embed{idx}'
-            t[col_name] = sentence_transformer_2(t.input, model_id=model_id, normalize_embeddings=True)
+            t[col_name] = sentence_transformer(t.input, model_id=model_id, normalize_embeddings=True)
             assert t.column_types()[col_name] == ArrayType((None,), dtype=FloatType(), nullable=False)
             list_col_name = f'embed_list{idx}'
-            t[list_col_name] = sentence_transformer_list_2(t.input_list, model_id=model_id, normalize_embeddings=True)
+            t[list_col_name] = sentence_transformer_list(t.input_list, model_id=model_id, normalize_embeddings=True)
             assert t.column_types()[list_col_name] == JsonType(nullable=True)
 
         def verify_row(row: Dict[str, Any]) -> None:
@@ -198,14 +198,14 @@ class TestFunctions:
         assert status.num_excs == 0
 
         # run multiple models one at a time in order to exercise batching
-        from pixeltable.functions.huggingface import cross_encoder_2, cross_encoder_list_2
+        from pixeltable.functions.huggingface import cross_encoder, cross_encoder_list
         model_ids = ['cross-encoder/ms-marco-MiniLM-L-6-v2', 'cross-encoder/ms-marco-TinyBERT-L-2-v2']
         for idx, model_id in enumerate(model_ids):
             col_name = f'embed{idx}'
-            t[col_name] = cross_encoder_2(t.input, t.input, model_id=model_id)
+            t[col_name] = cross_encoder(t.input, t.input, model_id=model_id)
             assert t.column_types()[col_name] == FloatType(nullable=True)
             list_col_name = f'embed_list{idx}'
-            t[list_col_name] = cross_encoder_list_2(t.input, t.input_list, model_id=model_id)
+            t[list_col_name] = cross_encoder_list(t.input, t.input_list, model_id=model_id)
             assert t.column_types()[list_col_name] == JsonType(nullable=True)
 
         def verify_row(row: Dict[str, Any]) -> None:
